Play A 'r.oguelike' Game

Clears the console and starts a game of 'r.oguelike'. The user inputs a keypress to move the character and interact with the game.

Usage

start_game(max_turns = 25)

Arguments

max_turns

Numeric. How many turns? Default is 25.

Value

Nothing. Clears the console and prints to it with cat.

Details

Use the WASD keys to move up, left, down and right. Use the '1' key to eat an apple from your inventory. If your terminal supports the 'keypress' package, then you can use a single keypress as input (e.g. the up arrow key), otherwise you will have to type at the prompt and then press 'Enter'. Use has_keypress_support to see if 'keypress' is supported.

Symbols used in the game are as follows:

. floor tile

# wall

@ player (10 HP max, -1 HP attack damage)

$ gold (+1 to +3 G)

E enemy (3 HP max, -1 HP attack damage)

a apple (+1 HP)

Examples

if (FALSE) start_game()